Share some more code in runtime-core::sys
